requirements.txt
langchain
langchain-openai
langchain_community
langchain_core
python-dotenv
streamlit